requirements.txt
streamlit
scipy
plotly
numpy
shapely
pyproj
simplekml
streamlit-aggrid
exceptiongroup
keyBERT
geopy
streamlit-aggrid
st_clickable_images
streamlit-authenticator
geopandas